Move validations to the src folder
