+++ src/containers/guest-home-page/login-dialog/LoginDialog.js
@@ -8,7 +8,7 @@ import LoginForm from '~/containers/guest-home-page/login-form/LoginForm'
 import useForm from '~/hooks/use-form'
 import useConfirm from '~/hooks/use-confirm'
 import { ModalContext } from '~/context/modal-context'
-import { email, password } from '~/constants/validation/login'
+import { email, password } from '~/validations/login'
 import loginImg from '~/assets/img/login-dialog/login.png'
 import { login } from '~/containers/guest-home-page/constants'
 import { loginUser } from '~/redux/reducer'
